syncing exercise 1 and 2

+++ exercise-4-integrate/src/main/java/za/co/entelect/dojo/service/CardService.java
@@ -4,5 +4,5 @@ import za.co.entelect.dojo.Card;
 
 public interface CardService {
 
-    void withdrawMoney(Card aCard, double transactionAmount);
+    void withdrawMoney(Card aCard, long transactionAmount);
 }
